responder cleaned up
