+++ src/includes/server/http/HTTPResponder.hpp
@@ -6,11 +6,8 @@
 #define HTTPRESPONDER_HPP
 
 #include <config/blocks/ServerBlock.hpp>
+#include "server/global/GlobalLogger.hpp"
 #include "server/http/HTTPClient.hpp"
-#include "env/env.hpp"
-#include "utils/base64.hpp"
-#include <climits>
-#include <cstdlib>
 #include <sys/stat.h>
 
 #ifdef BUILD_APPLE
@@ -23,24 +20,23 @@
 namespace NotApache {
 	class HTTPResponder {
 	public:
-		static void		generateResponse(HTTPClient &client);
+		static void generateResponse(HTTPClient &client);
 
 		static void generateAssociatedResponse(HTTPClient &client);
-
-		static void serveFile(HTTPClient &client, config::ServerBlock &server, config::RouteBlock &route, const std::string &file, const std::string &rewrittenUrl);
-
-		static void handleError(HTTPClient &client, config::ServerBlock *server, int code, bool doErrorPage = true);
-		static void handleError(HTTPClient &client, config::ServerBlock *server, config::RouteBlock *route, int code, bool doErrorPage = true);
-		static void serveDirectory(HTTPClient &client, config::ServerBlock &server, config::RouteBlock &route, const struct ::stat &directoryStat, const std::string &dirPath);
-		static void	prepareFile(HTTPClient &client, config::ServerBlock &server, config::RouteBlock &route, const utils::Uri &file, int code = 200, bool shouldErrorFile = true);
-		static void	prepareFile(HTTPClient &client, config::ServerBlock &server, config::RouteBlock &route, const struct ::stat &buf, const utils::Uri &file, int code = 200);
-		static void	uploadFile(HTTPClient &client, config::ServerBlock &server, config::RouteBlock &route, const std::string &f);
-		static void	deleteFile(HTTPClient &client, config::ServerBlock &server, config::RouteBlock &route, const std::string &f);
-
-		static void handleProxy(HTTPClient &client, config::ServerBlock *server, config::RouteBlock *route);
-
+		static void handleError(HTTPClient &client, int code);
 		static bool checkCredentials(const std::vector<std::string>& authFile, const std::string& credentials);
-		static void runCGI(HTTPClient& client, config::RouteBlock &route, const std::string& cgi, const std::string &rewrittenUrl);
+
+		static void serveFile(HTTPClient &client);
+		static void serveDirectory(HTTPClient &client, const struct ::stat &directoryStat);
+
+		static void	prepareFile(HTTPClient &client, int code = 200);
+		static void prepareFile(HTTPClient &client, const struct ::stat &buf, int code);
+
+		static void	uploadFile(HTTPClient &client);
+		static void	deleteFile(HTTPClient &client);
+
+		static void handleProxy(HTTPClient &client);
+		static void runCGI(HTTPClient& client, const std::string& cgi);
 
 		class AuthHeader : public std::exception {
 			public:
