+++ src/server/http/CGIClass.cpp
@@ -7,16 +7,32 @@
 #include "utils/intToString.hpp"
 #include "server/http/HTTPParser.hpp"
 #include <signal.h>
+#include <unistd.h>
 #include <algorithm>
 
 using namespace NotApache;
 
-CgiClass::CgiClass() : _envp(), pid(), status(0), response(HTTPParseData::CGI_RESPONSE), hasExited(true) {}
+CgiClass::CgiClass() :
+	_envp(),
+	pipefd(),
+	bodyPipefd(),
+	body(false),
+	pid(),
+	status(0),
+	response(HTTPParseData::CGI_RESPONSE), hasExited(true)
+{
+	args = new char *[3]();
+	pipefd[0] = -1;
+	pipefd[1] = -1;
+	bodyPipefd[0] = -1;
+	bodyPipefd[1] = -1;
+}
 
 CgiClass::~CgiClass() {
 	if (!hasExited) {
 		::kill(pid, SIGKILL);
 	}
+	freeArgs();
 }
 
 void CgiClass::generateENV(HTTPClient& client, const utils::Uri& uri, const std::string &rewrittenUrl) {
@@ -74,4 +90,61 @@ void CgiClass::generateENV(HTTPClient& client, const utils::Uri& uri, const std:
 
 CGIenv::env &CgiClass::getEnvp() {
 	return _envp;
+}
+
+void CgiClass::closePipes(FD *pipefd0, FD *pipefd1, FD *bodyPipefd0, FD *bodyPipefd1) {
+	bool closeFail = false;
+	if (pipefd0) {
+		if (::close(*pipefd0) == -1)
+			closeFail = true;
+	}
+	if (pipefd1) {
+		if (::close(*pipefd1) == -1)
+			closeFail = true;
+	}
+	if (bodyPipefd0) {
+		if (::close(*bodyPipefd0) == -1)
+			closeFail = true;
+	}
+	if (bodyPipefd1) {
+		if (::close(*bodyPipefd1) == -1)
+			closeFail = true;
+	}
+	if (closeFail)
+		ERROR_THROW(CgiClass::CloseFail());
+}
+
+void CgiClass::freeArgs() {
+	if (args) {
+		delete [] args[0];
+		delete [] args[1];
+		delete [] args;
+	}
+}
+
+void CgiClass::closePipes(FD *pipefd0, FD *pipefd1, FD *bodyPipefd0, FD *bodyPipefd1, bool closeFail) {
+	if (pipefd0) {
+		if (::close(*pipefd0) == -1)
+			closeFail = true;
+	}
+	if (pipefd1) {
+		if (::close(*pipefd1) == -1)
+			closeFail = true;
+	}
+	if (bodyPipefd0) {
+		if (::close(*bodyPipefd0) == -1)
+			closeFail = true;
+	}
+	if (bodyPipefd1) {
+		if (::close(*bodyPipefd1) == -1)
+			closeFail = true;
+	}
+	if (closeFail)
+		ERROR_THROW(CgiClass::CloseFail());
+}
+
+void CgiClass::freeArgs() {
+	delete [] args[0];
+	delete [] args[1];
+	delete [] args;
 }
