+++ tests/basic/basic.conf
@@ -1,0 +1,6 @@
+{
+    port 8080
+
+    route {
+        location /
+        root ../
